+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- "Connecting to remote cluster" screen no longer appears on launch when using provisioned mode without manual peers.
+
 ## [0.1.6] - 2026-02-24
 
 ### Fixed